Test GraphRequestAction after buildFromSession failure
This test simulates the case where a user provides a lambda that
generates a GraphStatement (as with one of the `executeGraphFluent`
overloads), but that lambda throws an exception instead of returning
normally.

The session-chain-checking portion of the test is original, but the
trailing assertions related to logging are adapted from a very similar
block of existing code in CqlRequestActionSpec.